+++ tests/Support/Objects/ExamplesAuthorizeQuery.php
@@ -3,7 +3,6 @@
 declare(strict_types = 1);
 namespace Rebing\GraphQL\Tests\Support\Objects;
 
-use Closure;
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Facades\GraphQL;
@@ -38,7 +37,7 @@ class ExamplesAuthorizeQuery extends Query
     /**
      * @param array<string,mixed> $args
      */
-    public function resolve(mixed $root, array $args, mixed $context, ResolveInfo $resolveInfo, Closure $getSelectFields): mixed
+    public function resolve(mixed $root, array $args, mixed $context, ResolveInfo $resolveInfo): mixed
     {
         $data = include __DIR__ . '/data.php';